+++ mitra_cli/src/main.rs
@@ -23,7 +23,7 @@ async fn main() {
         SubCommand::GenerateRsaKey(cmd) => cmd.execute().unwrap(),
         subcmd => {
             // Other commands require initialized app
-            let mut config = initialize_app(opts.log_level);
+            let mut config = initialize_app(Some(opts.log_level));
 
             let db_config = config.database_url.parse()
                 .expect("failed to parse database URL");